+++ src/agent/platformer_env.py
@@ -1,7 +1,7 @@
 import gym
 import numpy as np
 from gym import spaces
-from engine import GameEngine
+from game.engine import GameEngine
 
 class PlatformerEnv(gym.Env):
     """
