+++ src/tests_tiles.py
@@ -1,25 +1,25 @@
 import pygame
-from tiles import Block
-from tiles import Goal
-from tiles import Spike
+from game.tiles import Block
+from game.tiles import Goal
+from game.tiles import Spike
 import unittest
 
 class TestTiles(unittest.TestCase):
 
-	def test_update_block(self):
-		'''Check wether the update method work on blocks'''
+        def test_update_block(self):
+                '''Check whether the update method works on blocks'''
 		block=Block((20,20),1)
 		block.update(x_shift=10)
 		self.assertEqual(30,block.rect.x)
 
-	def test_update_spike(self):
-		'''Check wether the update method work on spikes'''
+        def test_update_spike(self):
+                '''Check whether the update method works on spikes'''
 		spike=Spike((20,20),1)
 		spike.update(x_shift=10)
 		self.assertEqual(30,spike.rect.x)
 
-	def test_update_goal(self):
-		'''Check wether the update method work on the goal'''
+        def test_update_goal(self):
+                '''Check whether the update method works on the goal'''
 		goal=Goal((20,20),1)
 		goal.update(x_shift=10)
 		self.assertEqual(30,goal.rect.x)
